pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dcsmm"
version = "0.6.0"
description = "Fast Upper Envelope Scan (Shanker & Dobrescu, 2023)."
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
  "numpy>=1.23",
  "numba",
  "scipy",
  "dill>=0.3.6",
  "econ-ark",
  "interpolation>=0.5.2",
  "ConSav",
  "pykdtree>=1.3.12",
]

[project.optional-dependencies]
examples = ["pyyaml", "matplotlib", "seaborn"]
mpi = ["mpi4py>=3.1"]

[project.urls]
"Source" = "https://github.com/akshayshanker/FUES"
"Bug Tracker" = "https://github.com/akshayshanker/FUES/issues"

[tool.setuptools.packages.find]
where = ["src"]
include = ["dcsmm*"]
